+++ apps/api/src/gift-cards/gift-cards.module.ts
@@ -1,5 +1,6 @@
 import { Module } from "@nestjs/common";
 import { StorefrontModule } from "../storefront/storefront.module";
+import { PlansModule } from "../plans/plans.module";
 import { RateLimitService } from "../common/rate-limit/rate-limit.service";
 import { SettingsModule } from "../settings-registry/settings.module";
 import { GiftCardsController } from "./gift-cards.controller";
@@ -7,7 +8,7 @@ import { GiftCardsService } from "./gift-cards.service";
 import { StorefrontGiftCardsController } from "./storefront-gift-cards.controller";
 
 @Module({
-  imports: [StorefrontModule, SettingsModule],
+  imports: [StorefrontModule, SettingsModule, PlansModule],
   controllers: [GiftCardsController, StorefrontGiftCardsController],
   providers: [GiftCardsService, RateLimitService],
   exports: [GiftCardsService],